+++ backend/canisters/user_index/api/src/queries/diamond_membership_fees.rs
@@ -1,4 +1,3 @@
-#![allow(deprecated)]
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
@@ -16,7 +15,6 @@ pub enum Response {
 #[ts_export(user_index, diamond_membership_fees)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct DiamondMembershipFees {
-    pub token: types::Cryptocurrency,
     pub token_symbol: String,
     pub one_month: u64,
     pub three_months: u64,