+++ backend/canisters/user_index/api/src/updates/pay_for_diamond_membership.rs
@@ -1,49 +1,16 @@
-#![allow(deprecated)]
 use candid::CandidType;
-use constants::{CHAT_LEDGER_CANISTER_ID, ICP_LEDGER_CANISTER_ID};
 use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{CanisterId, Cryptocurrency, DiamondMembershipPlanDuration, DiamondMembershipSubscription, TimestampMillis};
+use types::{CanisterId, DiamondMembershipPlanDuration, DiamondMembershipSubscription, TimestampMillis};
 
 #[ts_export(user_index, pay_for_diamond_membership)]
 #[derive(Serialize, Deserialize, Debug)]
-#[serde(from = "ArgsCombined")]
 pub struct Args {
     pub duration: DiamondMembershipPlanDuration,
-    pub token: Cryptocurrency,
     pub ledger: CanisterId,
     pub expected_price_e8s: u64,
     pub recurring: bool,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-struct ArgsCombined {
-    duration: DiamondMembershipPlanDuration,
-    token: Option<Cryptocurrency>,
-    ledger: Option<CanisterId>,
-    expected_price_e8s: u64,
-    recurring: bool,
-}
-
-impl From<ArgsCombined> for Args {
-    fn from(value: ArgsCombined) -> Self {
-        let token = value
-            .token
-            .unwrap_or(if matches!(value.ledger, Some(CHAT_LEDGER_CANISTER_ID)) {
-                Cryptocurrency::CHAT
-            } else {
-                Cryptocurrency::InternetComputer
-            });
-
-        Args {
-            duration: value.duration,
-            ledger: if matches!(token, Cryptocurrency::CHAT) { CHAT_LEDGER_CANISTER_ID } else { ICP_LEDGER_CANISTER_ID },
-            token,
-            expected_price_e8s: value.expected_price_e8s,
-            recurring: value.recurring,
-        }
-    }
 }
 
 #[ts_export(user_index, pay_for_diamond_membership)]